Remove uncertain rows from dataframes
some genders in the original dataframe are declared with "?". These are
ambivalent streetnames where the gender cant be assumed because its a
family name or the original authors werent sure what gender it is.
These cases cant be adressed with our approach (see #4) but for
benchmarking we will remove them and only compare the "sure" ones..
These are ambivalent streetnames where the gender cant be assumed
because its a family name or the original authors werent sure what
gender it is.#5
These cases cant be adressed with our approach (see #4) but for
benchmarking we will remove them and only compare the "sure" ones.